Added a new Project inside the soluction to house the EF Data Model of my database.
Added an ADO.net data entity model, named "MyDatabaseModel" with the username, password and database name to connect to my database.

Selected the checkbox that allows my username and password to go into the connectionstring.

Left all the options and settings in the wizard to what they were originally ie, did not change anything.